emotion_recognition_api/requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cycler==0.11.0
fastapi==0.73.0
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.13.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
numpy==1.22.2
oauthlib==3.2.0
opencv-python-headless
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorflow-gpu
termcolor==1.1.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
zipp==3.7.0